docs(README.md): update Map Component section to reflect the change from Mapbox GL library to Maplibre GL library
The Map Component section in the README.md file was updated to reflect the change from using the Mapbox GL library to using the Maplibre GL library. The link to the CSS file was also updated accordingly. This change was made to ensure accurate documentation and to provide the correct information to users.