fix(geo.ts): update condition to filter features based on date range
The condition to filter features based on the date range has been updated. Previously, it was checking if the feature had a property called "type" with a value of "poi". Now, it checks if the feature has properties called "metadata" with "dateStart" and "dateEnd" properties. If the feature has a date range and it falls within the current date, it will be returned. If the feature does not have a date range or it falls outside the current date, it will be filtered out.