feat(map): add support for using a proxy for raster tile URLs
The `Options` interface in `main.ts` now includes a new property `useProxy` which is a boolean indicating whether to use a proxy for raster tile URLs. If `useProxy` is set to `true`, the tile URLs will be modified to include the proxy URL. This allows for fetching raster tiles through a proxy server.

In the `Map` class, the `tiles` property of the `raster-tiles` source has been updated to conditionally include the proxy URL based on the value of `useProxy`. If `useProxy` is `true`, the proxy URL is prepended to the tile URL. If `useProxy` is `false`, the original tile URL is used.

This change improves flexibility and allows for using a proxy server to fetch raster tiles when necessary.